more, moving the type-directed expr generation code
